# api/core/v1beta1/openapi.yaml
openapi: 3.0.1
info:
  title: Flight Control API
  version: v1beta1
  description: |
    [Flight Control](https://flightctl.io) is a service for declarative management of fleets of edge devices and their workloads.
  contact:
    name: The Flight Control Team
    url: https://flightctl.io
    email: team@flightctl.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
...
paths:
...
  /auth/validate:
    get:
      tags:
        - authentication
      description: Validate an authentication token.
      operationId: authValidate
      parameters:
        - name: Authorization
          in: header
          description: The authentication token to validate.
          schema:
            type: string
      responses:
        "200":
          description: Token valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "418":
          description: Auth not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    orgId:
      type: apiKey
      in: query
      name: org_id
      description: The UUID of the organization that owns the resource.  Maps to an organizations metadata.name.
  schemas:
...
    Status:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        code:
          type: integer
          format: int32
          description: Suggested HTTP return code for this status, 0 if not set.
        message:
          type: string
          description: A human-readable description of the status of this operation.
        reason:
          type: string
          description: A machine-readable description of why this operation is in the "Failure" status. If this value is empty there is no information available. A Reason clarifies an HTTP status code but does not override it.
        status:
          type: string
          description: 'Status of the operation. One of: "Success" or "Failure". More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.'
      description: Status is a return value for calls that don't return other objects.
      required:
        - apiVersion
        - kind
        - status
        - code
        - reason
        - message
...